requirements.txt
PySide6>=6.5.0
Pillow>=9.0.0
psutil
pillow